+++ tests/test_future/test_future_container/test_future_equality.py
@@ -9,7 +9,7 @@ def test_nonequality():
     assert hash(Future.from_value(1))
 
 
-@pytest.mark.anyio
+@pytest.mark.anyio()
 async def test_equality():
     """Ensures that containers are not compared to regular values."""
     assert await Future.from_value(2) == await Future.from_value(2)